+++ errormessage/src/main/kotlin/motif/errormessage/InvalidQualifierHandler.kt
@@ -25,7 +25,7 @@ internal class InvalidQualifierHandler(private val error: InvalidQualifier) : Er
         appendln("""
             Qualifier must define either no members or a single value member of type String:
 
-              ${error.annotationType.qualifiedName}
+              ${error.annotation.className}
         """.trimIndent())
     }
 }