+++ models/src/main/kotlin/motif/models/ParsingError.kt
@@ -36,5 +36,5 @@ class NotAssignableBindsMethod(val objects: Objects, val method: IrMethod, val r
 class VoidDependenciesMethod(val scope: Scope, val dependenciesClass: IrClass, val method: IrMethod) : ParsingError()
 class DependencyMethodWithParameters(val scope: Scope, val dependenciesClass: IrClass, val method: IrMethod) : ParsingError()
 class NullableSpreadMethod(val objects: Objects, val factoryMethod: IrMethod, val spreadClass: IrClass, val spreadMethod: IrMethod) : ParsingError()
-class InvalidQualifier(val annotated: IrAnnotated, val annotationType: IrType) : ParsingError()
+class InvalidQualifier(val annotated: IrAnnotated, val annotation: IrAnnotation) : ParsingError()
 class DuplicatedChildParameterSource(val scope: Scope, val childScopeMethod: ChildMethod, val duplicatedParameters: List<ChildMethod.Parameter>) : ParsingError()